# docs/openapi.yaml
openapi: 3.1.0
info:
  title: RepairFlow API
  version: 0.1.0
  description: Contract for the RepairFlow MVP. Staff endpoints require a bearer token and validated X-Shop-Id tenant selection.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
...
paths:
...
  /api/v1/customers:
    get:
      tags: [Customers]
      summary: Search customers in the active shop
      operationId: listCustomers
      parameters:
        - $ref: '#/components/parameters/ShopHeader'
        - $ref: '#/components/parameters/Cursor'
        - name: query
          in: query
          schema: {type: string, maxLength: 100}
      responses:
        '200':
          description: Customer page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerListResponse'
        '403': {$ref: '#/components/responses/Forbidden'}
    post:
      tags: [Customers]
      summary: Create a customer
      operationId: createCustomer
      parameters:
        - $ref: '#/components/parameters/ShopHeader'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateCustomerRequest'}
      responses:
        '201':
          description: Customer created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CustomerResponse'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /api/v1/customers/{customerId}:
    patch:
      tags: [Customers]
      summary: Update an active customer
      operationId: updateCustomer
      parameters:
        - $ref: '#/components/parameters/ShopHeader'
        - $ref: '#/components/parameters/CustomerId'
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/UpdateCustomerRequest'}}}
      responses:
        '200': {description: Customer updated, content: {application/json: {schema: {$ref: '#/components/schemas/CustomerResponse'}}}}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/ValidationError'}
    delete:
      tags: [Customers]
      summary: Archive a customer and active devices
      operationId: archiveCustomer
      parameters:
        - $ref: '#/components/parameters/ShopHeader'
        - $ref: '#/components/parameters/CustomerId'
      responses:
        '200': {description: Customer archived, content: {application/json: {schema: {$ref: '#/components/schemas/CustomerResponse'}}}}
        '404': {$ref: '#/components/responses/NotFound'}
...
  /api/v1/media/presign:
    post:
      tags: [Repair orders]
      summary: Create a short-lived unbound upload for intake media
      operationId: presignIntakeMedia
      parameters:
        - $ref: '#/components/parameters/ShopHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PresignIntakeMediaRequest'}
      responses:
        '201':
          description: Signed upload for a provisional media asset
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PresignMediaResponse'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '422': {$ref: '#/components/responses/ValidationError'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
...
  /api/v1/repair-orders/{repairOrderId}/media/presign:
    post:
      tags: [Repair orders]
      summary: Create a short-lived signed upload request
      operationId: presignOrderMedia
      parameters:
        - $ref: '#/components/parameters/ShopHeader'
        - $ref: '#/components/parameters/RepairOrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PresignMediaRequest'}
      responses:
        '201':
          description: Signed upload
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PresignMediaResponse'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /api/v1/repair-orders/{repairOrderId}/media/{mediaAssetId}/download:
    get:
      tags: [Repair orders]
      summary: Create a short-lived signed download URL for private media
      operationId: downloadRepairOrderMedia
      parameters:
        - $ref: '#/components/parameters/ShopHeader'
        - $ref: '#/components/parameters/RepairOrderId'
        - name: mediaAssetId
          in: path
          required: true
          schema: {$ref: '#/components/schemas/Uuid'}
      responses:
        '200': {description: Signed private download, content: {application/json: {schema: {$ref: '#/components/schemas/DownloadMediaResponse'}}}}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: repairflow_refresh
  parameters:
    ShopHeader:
      name: X-Shop-Id
      in: header
      required: true
      description: Shop selector validated against the authenticated user's active membership.
      schema: {$ref: '#/components/schemas/Uuid'}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 16, maxLength: 128}
    Cursor:
      name: cursor
      in: query
      schema: {type: string}
    CustomerId:
      name: customerId
      in: path
      required: true
      schema: {$ref: '#/components/schemas/Uuid'}
    RepairOrderId:
      name: repairOrderId
      in: path
      required: true
      schema: {$ref: '#/components/schemas/Uuid'}
...
  responses:
...
    Forbidden:
      description: Authenticated actor lacks permission
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    NotFound:
      description: Resource not found in the active tenant
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
...
    ValidationError:
      description: Request validation failed
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
...
    ServiceUnavailable:
      description: A required external service is temporarily unavailable
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
...
  schemas:
...
    Uuid:
      type: string
      format: uuid
    IsoDateTime:
      type: string
      format: date-time
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: {type: string}
            message: {type: string}
            requestId: {type: string}
            details:
              type: array
              items:
                type: object
                required: [field, code]
                properties:
                  field: {type: string}
                  code: {type: string}
                  message: {type: string}
...
    CreateCustomerRequest:
      type: object
      additionalProperties: false
      required: [name, phone]
      properties:
        name: {type: string, minLength: 1, maxLength: 150}
        phone: {type: string, minLength: 8, maxLength: 30}
        email: {type: [string, 'null'], format: email}
        notes: {type: [string, 'null'], maxLength: 2000}
    UpdateCustomerRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name: {type: string, minLength: 1, maxLength: 150}
        phone: {type: string, minLength: 8, maxLength: 30}
        email: {type: [string, 'null'], format: email}
        notes: {type: [string, 'null'], maxLength: 2000}
    Customer:
      type: object
      required: [id, name, phone, createdAt, updatedAt]
      properties:
        id: {$ref: '#/components/schemas/Uuid'}
        name: {type: string}
        phone: {type: string}
        email: {type: [string, 'null']}
        notes: {type: [string, 'null']}
        createdAt: {$ref: '#/components/schemas/IsoDateTime'}
        updatedAt: {$ref: '#/components/schemas/IsoDateTime'}
    CustomerResponse:
      type: object
      required: [data]
      properties: {data: {$ref: '#/components/schemas/Customer'}}
    CustomerListResponse:
      type: object
      required: [data, meta]
      properties:
        data: {type: array, items: {$ref: '#/components/schemas/Customer'}}
        meta:
          type: object
          properties: {nextCursor: {type: [string, 'null']}}
...
    PresignMediaRequest:
      type: object
      additionalProperties: false
      required: [purpose, originalName, mimeType, byteSize]
      properties:
        purpose: {type: string, enum: [INTAKE, DIAGNOSIS, REPAIR, QC, HANDOVER, SIGNATURE]}
        originalName: {type: string, minLength: 1, maxLength: 255}
        mimeType: {type: string, enum: [image/jpeg, image/png, image/webp]}
        byteSize: {type: integer, minimum: 1, maximum: 15000000}
        checksumSha256: {type: [string, 'null'], pattern: '^[a-fA-F0-9]{64}$'}
    PresignIntakeMediaRequest:
      oneOf:
        - allOf:
            - {$ref: '#/components/schemas/PresignMediaRequest'}
            - type: object
              properties:
                purpose: {type: string, const: INTAKE}
        - type: object
          additionalProperties: false
          required: [purpose, originalName, mimeType, byteSize]
          properties:
            purpose: {type: string, const: AI_INTAKE_AUDIO}
            originalName: {type: string, minLength: 1, maxLength: 255}
            mimeType: {type: string, enum: [audio/wav, audio/x-wav]}
            byteSize: {type: integer, minimum: 1, maximum: 25000000}
            checksumSha256: {type: [string, 'null'], pattern: '^[a-fA-F0-9]{64}$'}
          description: Available only when the separately gated transcription boundary is configured.
    PresignMediaResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [mediaAssetId, uploadUrl, expiresAt]
          properties:
            mediaAssetId: {$ref: '#/components/schemas/Uuid'}
            uploadUrl: {type: string, format: uri}
            expiresAt: {$ref: '#/components/schemas/IsoDateTime'}
    DownloadMediaResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [downloadUrl, expiresAt, mimeType, originalName]
          properties:
            downloadUrl: {type: string, format: uri}
            expiresAt: {$ref: '#/components/schemas/IsoDateTime'}
            mimeType: {type: string}
            originalName: {type: string}
...
security:
  - bearerAuth: []